move metrics spec out from mp_continuous py
