+++ docs/source/mp/configuration.rst
@@ -561,25 +561,6 @@ Environment Variables
      - Set to ``1`` to disable anonymous usage statistics (cross-tool
        convention).
 
-Anonymous Usage Statistics
---------------------------
-
-At startup the MP server sends a one-time anonymous usage report to the
-LMCache stats server: an environment snapshot (cloud provider, CPU/GPU,
-memory) and a configuration snapshot (chunk size, transfer mode, L1
-size/medium, L2 adapter types, policies). While running, it also sends
-anonymous interval counters (tokens retrieved/stored and bytes stored in
-the last interval, plus uptime) every ``LMCACHE_USAGE_TRACK_INTERVAL``
-seconds (default 600). No prompts, keys, KV-cache data, model names, or
-the server's ``--instance-id`` are ever sent; reports are correlated only
-through random UUIDs. Reporting runs on background threads, never delays
-startup or request handling, and a failure anywhere in reporting never
-affects serving.
-
-To opt out, set ``LMCACHE_TRACK_USAGE=false`` or ``DO_NOT_TRACK=1``. See
-:ref:`usage-stats-collection` for the full list of collected fields and
-configuration options.
-
 Full Example
 ------------
 
@@ -615,3 +596,14 @@ Full Example
         --metrics-sample-rate 0.01 \
         --enable-tracing \
         --otlp-endpoint http://localhost:4317
+
+Anonymous Usage Statistics
+--------------------------
+
+The MP server reports anonymous usage statistics: a one-time
+environment/configuration snapshot at startup and interval counters
+(tokens retrieved/stored, bytes stored, uptime) every
+``LMCACHE_USAGE_TRACK_INTERVAL`` seconds. No prompts, keys, KV-cache
+data, model names, or ``--instance-id`` are ever sent, and reporting can
+never affect serving. Opt out with ``LMCACHE_TRACK_USAGE=false`` or
+``DO_NOT_TRACK=1``; see :ref:`usage-stats-collection` for details.
